style: move cover plugin to wrapper folder
