+++ packages/website/astro.config.ts
@@ -1,7 +1,7 @@
 import react from '@astrojs/react';
 import sitemap from '@astrojs/sitemap';
 import { defineConfig } from 'astro/config';
-import { coverPlugin } from './remark-cover-plugin.ts';
+import { coverPlugin } from './markdown-plugins/remark-cover/index.ts';
 
 // https://astro.build/config
 export default defineConfig({
